created files for common functions
